backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic[email]==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
pyotp==2.9.0
qrcode[pil]==7.4.2
redis==5.0.1
slowapi==0.1.9
python-dotenv==1.0.0
httpx==0.26.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==22.0.0
